Tests and bug fixes for 'elbow' nodes and tree label search.
'Elbow' nodes are nodes with exactly one parent and one child. The Newick tree

'((a,(b)c)d,e);'

for example, has an elbow node labeled 'c' with a child node labeled
'b'. Ape currently parses trees with these nodes incorrectly, and
ggphylo's tree.read function includes a simple fix for this.

Second, the tree.label.for.node function was updated to return a
vector of labels if given a vector of nodes on input. In general, when
possible all tree functions should accept vector input/output. More
work needs to be done on this.