Fix up docs and add gridExtra dependency.
This commit makes some changes to the Roxygen docs (namely, adding
S3method annotations) which allow 'R CMD check' to now be run without
errors (on my system at least...).

It also adds a missing dependency on the gridExtra package, which is
used in the vignette. Hopefully one of these changes fixes the
installation problems encountered in issue #4.